support add/modify-only import job (Bug #41350)
git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.1/ucs-school-4.1r2@69561 1b283449-9f2f-0410-b571-8eb228e1a901

+++ ucs-school-import/modules/ucsschool/importer/mass_import/mass_import.py
@@ -60,9 +60,12 @@ class MassImport(object):
 
 	def import_users(self):
 		user_import = self.factory.make_user_importer(self.dry_run)
-		imported_users = user_import.import_users()
+		imported_users = user_import.read_input()
 		user_import.create_and_modify_users(imported_users)
-		users_to_delete = user_import.detect_users_to_delete()
-		user_import.delete_users(users_to_delete)
+		if self.config["no_delete"]:
+			self.logger.info("------ Skipping user deletion (no_delete=%r) ------", self.config["no_delete"])
+		else:
+			users_to_delete = user_import.detect_users_to_delete()
+			user_import.delete_users(users_to_delete)
 		user_import.log_stats()
 		self.writer.output(imported_users, user_import.deleted_users)
